LPS-165682 Move ObjectCodeEditorUtil to object-api for exporting
